+++ src/components/common/form/input-string.tsx
@@ -2,15 +2,17 @@
 
 import TextField from '@mui/material/TextField';
 import { ChangeEvent } from 'react';
+import { FieldError } from 'react-hook-form';
 
 interface InputStringProps {
     label: string;
     variant?: 'outlined' | 'filled' | 'standard';
     value: string;
     setValue: (newValue: string) => void;
+    error?: FieldError;
 }
 
-export function InputString({ label, variant = 'standard', value, setValue }: InputStringProps) {
+export function InputString({ label, variant = 'standard', value, setValue, error }: InputStringProps) {
     const onChange = (event: ChangeEvent<HTMLInputElement>) => {
         setValue(event.target.value);
     };
@@ -21,6 +23,10 @@ export function InputString({ label, variant = 'standard', value, setValue }: In
             variant={variant}
             value={value}
             onChange={onChange}
+            error={Boolean(error)}
+            helperText={error?.message ?? ''}
+            autoComplete="new-password"
+            type="text"
         />
     );
 }