+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.28.0] - 2024-08-17
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `react-hook-form` library.
+- [Marek Kurańda](https://github.com/mjkuranda): error messaging in `InputString` and `InputPassword`.
+- [Marek Kurańda](https://github.com/mjkuranda): user registering capability at `users/registration`.
+
+### Fixed
+- [Marek Kurańda](https://github.com/mjkuranda): toasts visibility.
+
 ## [1.27.0] - 2024-08-16
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `Loader` component.